# [Enhancement] Optimize avro reader performance