--- a/be/src/formats/avro/cpp/avro_reader.cpp
+++ b/be/src/formats/avro/cpp/avro_reader.cpp
@@ -20,6 +20,7 @@
 #include <avrocpp/Types.hh>
 #include <avrocpp/ValidSchema.hh>
 
+#include "column/adaptive_nullable_column.h"
 #include "exec/file_scanner.h"
 #include "formats/avro/cpp/avro_schema_builder.h"
 #include "formats/avro/cpp/utils.h"
@@ -103,21 +104,44 @@ AvroReader::~AvroReader() {
 Status AvroReader::init(std::unique_ptr<avro::InputStream> input_stream, const std::string& filename,
                         RuntimeState* state, ScannerCounter* counter, const std::vector<SlotDescriptor*>* slot_descs,
                         const std::vector<avrocpp::ColumnReaderUniquePtr>* column_readers, bool col_not_found_as_null) {
+    if (_is_inited) {
+        return Status::OK();
+    }
+
+    _filename = filename;
+    _slot_descs = slot_descs;
+    _column_readers = column_readers;
+    _num_of_columns_from_file = _column_readers->size();
+    _col_not_found_as_null = col_not_found_as_null;
+
+    _state = state;
+    _counter = counter;
+
     try {
         _file_reader = std::make_unique<avro::DataFileReader<avro::GenericDatum>>(std::move(input_stream));
-        _datum = std::make_unique<avro::GenericDatum>(_file_reader->dataSchema());
+
+        const auto& schema = _file_reader->dataSchema();
+        _datum = std::make_unique<avro::GenericDatum>(schema);
+
+        _field_indexes.resize(_num_of_columns_from_file, -1);
+        for (size_t i = 0; i < _num_of_columns_from_file; ++i) {
+            const auto& desc = (*_slot_descs)[i];
+            if (desc == nullptr) {
+                continue;
+            }
+
+            size_t index = 0;
+            if (schema.root()->nameIndex(desc->col_name(), index)) {
+                _field_indexes[i] = index;
+            }
+        }
     } catch (const avro::Exception& ex) {
         auto err_msg = fmt::format("Avro reader init throws exception: {}", ex.what());
         LOG(WARNING) << err_msg;
         return Status::InternalError(err_msg);
     }
 
-    _filename = filename;
-    _state = state;
-    _counter = counter;
-    _slot_descs = slot_descs;
-    _column_readers = column_readers;
-    _col_not_found_as_null = col_not_found_as_null;
+    _is_inited = true;
     return Status::OK();
 }
 
@@ -126,12 +150,43 @@ void AvroReader::TEST_init(const std::vector<SlotDescriptor*>* slot_descs,
                            bool col_not_found_as_null) {
     _slot_descs = slot_descs;
     _column_readers = column_readers;
+    _num_of_columns_from_file = _column_readers->size();
     _col_not_found_as_null = col_not_found_as_null;
+
+    const auto& schema = _file_reader->dataSchema();
+    _datum = std::make_unique<avro::GenericDatum>(schema);
+
+    _field_indexes.resize(_num_of_columns_from_file, -1);
+    for (size_t i = 0; i < _num_of_columns_from_file; ++i) {
+        const auto& desc = (*_slot_descs)[i];
+        if (desc == nullptr) {
+            continue;
+        }
+
+        size_t index = 0;
+        if (schema.root()->nameIndex(desc->col_name(), index)) {
+            _field_indexes[i] = index;
+        }
+    }
+
+    _is_inited = true;
 }
 
 Status AvroReader::read_chunk(ChunkPtr& chunk, int rows_to_read) {
-    if (_file_reader == nullptr || _column_readers == nullptr) {
+    if (!_is_inited) {
         return Status::Uninitialized("Avro reader is not initialized");
+    }
+
+    // get column raw ptrs before reading chunk
+    std::vector<AdaptiveNullableColumn*> column_raw_ptrs;
+    column_raw_ptrs.resize(_num_of_columns_from_file, nullptr);
+    for (size_t i = 0; i < _num_of_columns_from_file; ++i) {
+        const auto& desc = (*_slot_descs)[i];
+        if (desc == nullptr) {
+            continue;
+        }
+
+        column_raw_ptrs[i] = down_cast<AdaptiveNullableColumn*>(chunk->get_column_by_slot_id(desc->id()).get());
     }
 
     try {
@@ -141,7 +196,7 @@ Status AvroReader::read_chunk(ChunkPtr& chunk, int rows_to_read) {
             DCHECK(_datum->type() == avro::AVRO_RECORD);
             const auto& record = _datum->value<avro::GenericRecord>();
 
-            auto st = read_row(chunk, record);
+            auto st = read_row(record, column_raw_ptrs);
             if (st.is_data_quality_error()) {
                 if (_counter->num_rows_filtered++ < MAX_ERROR_LINES_IN_FILE) {
                     std::string json_str;
@@ -171,34 +226,34 @@ Status AvroReader::read_chunk(ChunkPtr& chunk, int rows_to_read) {
     }
 }
 
-Status AvroReader::read_row(ChunkPtr& chunk, const avro::GenericRecord& record) {
-    size_t num_of_columns_from_file = _column_readers->size();
-    for (size_t i = 0; i < num_of_columns_from_file; ++i) {
+Status AvroReader::read_row(const avro::GenericRecord& record,
+                            const std::vector<AdaptiveNullableColumn*>& column_raw_ptrs) {
+    for (size_t i = 0; i < _num_of_columns_from_file; ++i) {
         const auto& desc = (*_slot_descs)[i];
         if (desc == nullptr) {
             continue;
         }
 
-        auto* column = down_cast<NullableColumn*>(chunk->get_column_by_slot_id(desc->id()).get());
-        auto col_name = desc->col_name();
-        if (record.hasField(col_name)) {
-            const auto& field = record.field(col_name);
+        DCHECK(column_raw_ptrs[i] != nullptr);
+
+        if (_field_indexes[i] >= 0) {
             DCHECK((*_column_readers)[i] != nullptr);
-            RETURN_IF_ERROR((*_column_readers)[i]->read_datum_for_adaptive_column(field, column));
+            const auto& field = record.fieldAt(_field_indexes[i]);
+            RETURN_IF_ERROR((*_column_readers)[i]->read_datum_for_adaptive_column(field, column_raw_ptrs[i]));
         } else if (!_col_not_found_as_null) {
             return Status::NotFound(
                     fmt::format("Column: {} is not found in file: {}. Consider setting "
                                 "'fill_mismatch_column_with' = 'null' property",
-                                col_name, _filename));
+                                desc->col_name(), _filename));
         } else {
-            column->append_nulls(1);
+            column_raw_ptrs[i]->append_nulls(1);
         }
     }
     return Status::OK();
 }
 
 Status AvroReader::get_schema(std::vector<SlotDescriptor>* schema) {
-    if (_file_reader == nullptr) {
+    if (!_is_inited) {
         return Status::Uninitialized("Avro reader is not initialized");
     }