## Why I'm doing:

## What I'm doing:
1. generate field indexes when init reader. (`record.hasField(name)` and `record.field(name)` cost too much time)
2. reuse `GenericDatum` in reader.
3. generate column raw ptrs when read chunk start.

```
ssb lineorder 6001215 lines

SQL1: select count(lo_orderkey) from files("path" = "hdfs://hdfs/lineorder.avro", "format" = "avro");
4.62s -> 2.18s

SQL2: insert into blackhole() select * from files("path" = "hdfs://hdfs/lineorder.avro", "format" = "avro");
11.6s -> 5.36s
```

Fixes #58157 

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [x] 3.5
  - [x] 3.4
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1